+++ src/main/java/step3/GameResult.java
@@ -1,5 +1,8 @@
 package step3;
 
+import step3.domain.Cars;
+import step3.domain.Position;
+
 import java.util.List;
 
 public class GameResult {
