--- src/main/java/step3/Score.java
@@ -1,4 +1,0 @@
-package step3;
-
-public class Score {
-}
