+++ src/main/java/step3/domain/Cars.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 import step3.exception.IllegalCarsException;
