+++ src/main/java/step3/domain/GameConsole.java
@@ -1,4 +1,6 @@
-package step3;
+package step3.domain;
+
+import step3.GameResult;
 
 public class GameConsole {
