+++ src/main/java/step3/domain/RandomGenerator.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 import java.util.Random;
