+++ src/test/java/step3/CarsTest.java
@@ -2,6 +2,8 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Cars;
+import step3.domain.Position;
 import step3.exception.IllegalCarsException;
 
 import java.util.Arrays;
