+++ src/test/java/step3/ContestsTest.java
@@ -2,6 +2,9 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Cars;
+import step3.domain.Contests;
+import step3.domain.Position;
 import step3.exception.IllegalContestsException;
 
 import java.util.Arrays;
