+++ src/test/java/step3/MoveStratgyTest.java
@@ -3,6 +3,7 @@ package step3;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import step3.domain.MoveStratgy;
 
 import static org.assertj.core.api.Assertions.assertThat;
